Improved file loading
Eventhough I have implemented async functions and Promises, the `openDirectory` was still blocking, because I used synchronous functions to read all items of a directory and reading file stats. The solution was to wrap a Promise around `fs.readdir(...)` and `fs.stat(...)`, so that these functions can be called with the `await` modifier.