Un-ignored PNG files
I intented to ignore the PNG files since they can be generating using the provided script, but the app depends on these scripts (errors occur if electron is unable to find the icons) and the generation script is not part of the bundling process. The script could have been run by bundle.js but that's counter-intuitive for the these static files. So the PNG files of different sizes were added to `app/images/`